requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.11
python-docx==1.1.0
python-dotenv==1.0.0
tqdm==4.66.1
orjson==3.9.0
watchdog==3.0.0
